requirements.txt
base58==0.2.2
boto3==1.3.1
botocore==1.4.13
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-autocomplete-light==3.3.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-filter==2.0.0
django-smart-selects==1.5.4
django-zappa==0.13.1
djangorestframework==3.9.0
docopt==0.6.2
docutils==0.12
futures==3.0.5
hjson==1.5.4
icalendar==4.0.3
idna==2.8
jmespath==0.9.0
lambda-packages==0.5.0
mysql-connector-python-rf==2.1.3
Pillow==4.1.1
psycopg2==2.7.6.1
python-dateutil==2.5.3
python-slugify==1.2.0
pytz==2018.7
requests==2.21.0
shortuuid==0.5.0
six==1.10.0
sqlparse==0.2.4
tqdm==4.4.1
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.11.8
wsgi-request-logger==0.4.5
zappa==0.17.4
